fixed wrong references to old refactoring and uncorrect lib path
